running integration tests part of suite
