+++ src/it/scala/castalia/IntegrationTestBase.scala
@@ -1,21 +1,12 @@
 package castalia
 
-import castalia.utils.TestHelpers
-import org.scalatest.{BeforeAndAfterAll, FunSpec, GivenWhenThen, Matchers}
+import org.scalatest._
 
 
 /**
  * Created by jens on 06-12-15.
  */
-trait IntegrationTestBase extends FunSpec with GivenWhenThen with Matchers with BeforeAndAfterAll {
-   // TODO make configurable
-  val serverAddress = "localhost:9000"
-  val server = Main
+trait IntegrationTestBase extends FunSpec with GivenWhenThen with Matchers {
 
-  override def beforeAll(): Unit = {
-    if (!TestHelpers.isServerRunning(serverAddress)) {
-      println("server is not running... spinning up.")
-      server.main(Array("castaliaT.json"))
-    }
-  }
+  val serverAddress = "localhost:9000"
 }
