+++ src/it/scala/castalia/JsonConfiguredStubServerTest.scala
@@ -3,10 +3,12 @@ package castalia
 import com.twitter.finagle
 import com.twitter.finagle.http.{Method, Request}
 import com.twitter.util.Await
+import org.scalatest.DoNotDiscover
 
 /**
   * Created by rezolya on 05/12/2015.
   */
+@DoNotDiscover
 class JsonConfiguredStubServerTest extends IntegrationTestBase {
   // TODO: other way than starting application using empty string array in main?
